+++ tests/lightning_qubit/test_measurements_class.py
@@ -38,8 +38,6 @@ if not LightningDevice._new_API:
         allow_module_level=True,
     )
 
-if device_name == "lightning.tensor":
-    pytest.skip("Skipping tests for the LightningTensor class.", allow_module_level=True)
 
 if not LightningDevice._CPP_BINARY_AVAILABLE:
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
@@ -66,8 +64,11 @@ def test_initialization(lightning_sv):
     statevector = lightning_sv(num_wires=5)
     m = LightningMeasurements(statevector)
 
-    assert m.qubit_state is statevector
-    assert m.dtype == statevector.dtype
+    if device_name == "lightning.tensor":
+        assert m.dtype == statevector.dtype
+    else:
+        assert m.qubit_state is statevector
+        assert m.dtype == statevector.dtype
 
 
 class TestGetMeasurementFunction:
@@ -96,6 +97,9 @@ class TestGetMeasurementFunction:
     )
     def test_state_diagonalizing_gates_measurements(self, lightning_sv, mp):
         """Test that any non-expval measurement calls the state_diagonalizing_gates method"""
+        if isinstance(mp.obs, qml.Projector) and device_name == "lightning.tensor":
+            pytest.skip("lightning.tensor does not support Projector.")
+
         statevector = lightning_sv(num_wires=5)
         m = LightningMeasurements(statevector)
 
@@ -117,6 +121,8 @@ class TestGetMeasurementFunction:
     )
     def test_expval_selected(self, lightning_sv, obs):
         """Test that expval is chosen for a variety of different expectation values."""
+        if isinstance(obs, qml.SparseHamiltonian) and device_name == "lightning.tensor":
+            pytest.skip("lightning.tensor does not support SparseHamiltonian.")
         statevector = lightning_sv(num_wires=5)
         m = LightningMeasurements(statevector)
         mp = qml.expval(obs)
@@ -174,6 +180,10 @@ class TestStateDiagonalizingGates:
         result = getattr(m, method_name)(qml.expval(qml.I(4)))
         assert np.allclose(result, 1.0)
 
+    @pytest.mark.skipif(
+        device_name == "lightning.tensor",
+        reason="lightning.tensor does not support a single-wire circuit.",
+    )
     def test_basis_state_projector_expval(self, lightning_sv, method_name):
         """Test expectation value for a basis state projector."""
         phi = 0.8
@@ -183,6 +193,10 @@ class TestStateDiagonalizingGates:
         result = getattr(m, method_name)(qml.expval(qml.Projector([0], wires=0)))
         assert qml.math.allclose(result, np.cos(phi / 2) ** 2)
 
+    @pytest.mark.skipif(
+        device_name == "lightning.tensor",
+        reason="lightning.tensor does not support a single-wire circuit.",
+    )
     def test_state_vector_projector_expval(self, lightning_sv, method_name):
         """Test expectation value for a state vector projector."""
         phi = -0.6
@@ -212,9 +226,17 @@ class TestExpval:
         tape = qml.tape.QuantumScript(ops, measurements)
 
         statevector = lightning_sv(wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = np.cos(theta)
 
         assert np.allclose(result, expected, tol)
@@ -228,9 +250,17 @@ class TestExpval:
             [qml.expval(qml.Identity(wires=[0])), qml.expval(qml.Identity(wires=[1]))],
         )
         statevector = lightning_sv(wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = 1.0
 
         assert np.allclose(result, expected, tol)
@@ -243,9 +273,17 @@ class TestExpval:
             [qml.expval(qml.Identity(wires=[0, 1]))],
         )
         statevector = lightning_sv(wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = 1.0
 
         assert np.allclose(result, expected, tol)
@@ -291,9 +329,17 @@ class TestExpval:
             [qml.expval(Obs[0]), qml.expval(Obs[1])],
         )
         statevector = lightning_sv(wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = expected_fn(theta, phi)
 
         assert np.allclose(result, expected, tol)
@@ -330,6 +376,10 @@ class TestExpvalHamiltonian:
     )
     def test_expval_hamiltonian(self, obs, coeffs, expected, tol, lightning_sv, method_name):
         """Test expval with Hamiltonian"""
+
+        if any(isinstance(o, qml.Hermitian) for o in obs) and device_name == "lightning.tensor":
+            pytest.skip("Hermitian with 1+ wires target not supported in lightning.tensor.")
+
         ham = qml.Hamiltonian(coeffs, obs)
 
         statevector = lightning_sv(self.wires)
@@ -341,6 +391,9 @@ class TestExpvalHamiltonian:
         assert np.allclose(result, expected, atol=tol, rtol=0)
 
 
+@pytest.mark.skipif(
+    device_name == "lightning.tensor", reason="lightning.tensor does not support sparseH."
+)
 class TestSparseExpval:
     """Tests for the expval function"""
 
@@ -388,9 +441,7 @@ class TestMeasurements:
         for m in tape.measurements:
             # NotImplementedError in DefaultQubit
             # We therefore validate against `qml.Hermitian`
-            if isinstance(m, VarianceMP) and isinstance(
-                m.obs, (qml.Hamiltonian, qml.SparseHamiltonian)
-            ):
+            if isinstance(m, VarianceMP) and isinstance(m.obs, (qml.SparseHamiltonian)):
                 use_default = False
                 new_meas.append(m.__class__(qml.Hermitian(qml.matrix(m.obs), wires=m.obs.wires)))
                 continue
@@ -404,9 +455,17 @@ class TestMeasurements:
 
         tape = qml.tape.QuantumScript(tape.operations, new_meas)
         statevector = lightning_sv(tape.num_wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        return m.measure_final_state(tape)
+        return (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
 
     @flaky(max_runs=15)
     @pytest.mark.parametrize("shots", [None, 200_000, [190_000, 190_000]])
@@ -432,6 +491,17 @@ class TestMeasurements:
         ),
     )
     def test_single_return_value(self, shots, measurement, observable, lightning_sv, tol):
+        if (
+            isinstance(observable, qml.Hermitian)
+            and len(observable.wires) > 1
+            and device_name == "lightning.tensor"
+        ):
+            pytest.skip("lightning.tensor does not support 1+ wires Hermitian.")
+        if observable == [1, 0] and device_name == "lightning.tensor" and measurement is qml.probs:
+            pytest.skip("lightning.tensor does not support out of order probs.")
+        if isinstance(observable, qml.SparseHamiltonian) and device_name == "lightning.tensor":
+            pytest.skip("lightning.tensor does not support SparseHamiltonian.")
+
         if measurement is qml.probs and isinstance(
             observable,
             (
@@ -465,7 +535,8 @@ class TestMeasurements:
         np.random.seed(0)
         weights = np.random.rand(n_layers, n_qubits, 3)
         ops = [qml.Hadamard(i) for i in range(n_qubits)]
-        ops += [qml.StronglyEntanglingLayers(weights, wires=range(n_qubits))]
+        if device_name != "lightning.tensor":
+            ops += [qml.StronglyEntanglingLayers(weights, wires=range(n_qubits))]
         measurements = (
             [measurement(wires=observable)]
             if isinstance(observable, list)
@@ -474,7 +545,11 @@ class TestMeasurements:
         tape = qml.tape.QuantumScript(ops, measurements, shots=shots)
 
         statevector = lightning_sv(n_qubits)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
 
         skip_list = (
@@ -488,10 +563,18 @@ class TestMeasurements:
         do_skip = do_skip and shots is not None
         if do_skip:
             with pytest.raises(TypeError):
-                _ = m.measure_final_state(tape)
+                _ = (
+                    m.measure_final_state(tape)
+                    if device_name != "lightning.tensor"
+                    else m.measure_tensor_network(tape)
+                )
             return
         else:
-            result = m.measure_final_state(tape)
+            result = (
+                m.measure_final_state(tape)
+                if device_name != "lightning.tensor"
+                else m.measure_tensor_network(tape)
+            )
 
         expected = self.calculate_reference(tape, lightning_sv)
 
@@ -543,6 +626,18 @@ class TestMeasurements:
         ),
     )
     def test_double_return_value(self, shots, measurement, obs0_, obs1_, lightning_sv, tol):
+        if (
+            any(isinstance(obs, qml.Hermitian) for obs in [obs0_, obs1_])
+            and any(len(obs.wires) > 1 for obs in [obs0_, obs1_])
+            and device_name == "lightning.tensor"
+        ):
+            pytest.skip("lightning.tensor does not support 1+ wires Hermitian.")
+        if (
+            any(isinstance(obs, qml.SparseHamiltonian) for obs in [obs0_, obs1_])
+            and device_name == "lightning.tensor"
+        ):
+            pytest.skip("lightning.tensor does not support SparseHamiltonian.")
+
         skip_list = (
             qml.ops.Sum,
             qml.ops.SProd,
@@ -573,12 +668,17 @@ class TestMeasurements:
         np.random.seed(0)
         weights = np.random.rand(n_layers, n_qubits, 3)
         ops = [qml.Hadamard(i) for i in range(n_qubits)]
-        ops += [qml.StronglyEntanglingLayers(weights, wires=range(n_qubits))]
+        if device_name != "lightning.tensor":
+            ops += [qml.StronglyEntanglingLayers(weights, wires=range(n_qubits))]
         measurements = [measurement(op=obs0_), measurement(op=obs1_)]
         tape = qml.tape.QuantumScript(ops, measurements, shots=shots)
 
         statevector = lightning_sv(n_qubits)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
 
         skip_list = (
@@ -598,10 +698,18 @@ class TestMeasurements:
         do_skip = do_skip and shots is not None
         if do_skip:
             with pytest.raises(TypeError):
-                _ = m.measure_final_state(tape)
+                _ = (
+                    m.measure_final_state(tape)
+                    if device_name != "lightning.tensor"
+                    else m.measure_tensor_network(tape)
+                )
             return
         else:
-            result = m.measure_final_state(tape)
+            result = (
+                m.measure_final_state(tape)
+                if device_name != "lightning.tensor"
+                else m.measure_tensor_network(tape)
+            )
 
         expected = self.calculate_reference(tape, lightning_sv)
         if len(expected) == 1:
@@ -619,6 +727,10 @@ class TestMeasurements:
             # allclose -> absolute(r - e) <= (atol + rtol * absolute(e))
             assert np.allclose(r, e, atol=atol, rtol=rtol)
 
+    @pytest.mark.skipif(
+        device_name == "lightning.tensor",
+        reason="lightning.tensor does not support out of order probs.",
+    )
     @pytest.mark.parametrize(
         "cases",
         [
@@ -687,9 +799,19 @@ class TestControlledOps:
     @pytest.mark.parametrize("n_qubits", list(range(2, 5)))
     def test_controlled_qubit_gates(self, operation, n_qubits, control_value, tol, lightning_sv):
         """Test that multi-controlled gates are correctly applied to a state"""
-        threshold = 250
+        threshold = 250 if device_name != "lightning.tensor" else 5
         num_wires = max(operation.num_wires, 1)
         np.random.seed(0)
+
+        if num_wires > 1 and device_name == "lightning.tensor":
+            pytest.skip(
+                "Multi target wires control operations are not supported in lightning.tensor."
+            )
+
+        if (
+            operation == qml.GlobalPhase or operation == qml.MultiRZ
+        ) and device_name == "lightning.tensor":
+            pytest.skip("GlobalPhase and MultiRZ are not supported in lightning.tensor.")
 
         for n_wires in range(num_wires + 1, num_wires + 4):
             wire_lists = list(itertools.permutations(range(0, n_qubits), n_wires))
@@ -711,9 +833,11 @@ class TestControlledOps:
                         qml.ctrl(
                             operation(target_wires),
                             control_wires,
-                            control_values=[
-                                control_value or bool(i % 2) for i, _ in enumerate(control_wires)
-                            ],
+                            control_values=(
+                                [control_value or bool(i % 2) for i, _ in enumerate(control_wires)]
+                                if device_name != "lightning.tensor"
+                                else [control_value for _ in control_wires]
+                            ),
                         ),
                     ]
                 else:
@@ -721,9 +845,11 @@ class TestControlledOps:
                         qml.ctrl(
                             operation(*tuple([0.1234] * operation.num_params), target_wires),
                             control_wires,
-                            control_values=[
-                                control_value or bool(i % 2) for i, _ in enumerate(control_wires)
-                            ],
+                            control_values=(
+                                [control_value or bool(i % 2) for i, _ in enumerate(control_wires)]
+                                if device_name != "lightning.tensor"
+                                else [control_value for _ in control_wires]
+                            ),
                         ),
                     ]
 
@@ -731,13 +857,27 @@ class TestControlledOps:
                 tape = qml.tape.QuantumScript(ops, measurements)
 
                 statevector = lightning_sv(n_qubits)
-                statevector = statevector.get_final_state(tape)
+                statevector = (
+                    statevector.get_final_state(tape)
+                    if device_name != "lightning.tensor"
+                    else statevector.set_tensor_network(tape)
+                )
                 m = LightningMeasurements(statevector)
-                result = m.measure_final_state(tape)
+                result = (
+                    m.measure_final_state(tape)
+                    if device_name != "lightning.tensor"
+                    else m.measure_tensor_network(tape)
+                )
                 expected = self.calculate_reference(tape)
+                if device_name == "lightning.tensor":
+                    assert np.allclose(result, expected, 1e-4)
+                else:
+                    assert np.allclose(result, expected, tol * 10)
 
-                assert np.allclose(result, expected, tol * 10)
-
+    @pytest.mark.skipif(
+        device_name != "lightning.qubit",
+        reason="N-controlled operations only implemented in lightning.qubit.",
+    )
     def test_controlled_qubit_unitary_from_op(self, tol, lightning_sv):
         n_qubits = 10
         par = 0.1234
@@ -752,9 +892,17 @@ class TestControlledOps:
         )
 
         statevector = lightning_sv(n_qubits)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = self.calculate_reference(tape)
 
         assert np.allclose(result, expected, tol)
@@ -785,13 +933,27 @@ class TestControlledOps:
         )
 
         statevector = lightning_sv(n_qubits)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = self.calculate_reference(tape_cnot)
 
-        assert np.allclose(result, expected, tol)
+        if device_name == "lightning.tensor":
+            assert np.allclose(result, expected, 1e-4)
+        else:
+            assert np.allclose(result, expected, tol)
 
+    @pytest.mark.skipif(
+        device_name == "lightning.tensor", reason="lightning.tensor does not support GlobalPhase."
+    )
     @pytest.mark.parametrize("control_value", [False, True])
     @pytest.mark.parametrize("n_qubits", list(range(2, 8)))
     def test_controlled_globalphase(self, n_qubits, control_value, tol, lightning_sv):
@@ -845,9 +1007,17 @@ class TestExpOperatorArithmetic:
             [qml.expval(qml.s_prod(0.5, qml.PauliZ(0)))],
         )
         statevector = lightning_sv(self.wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = 0.5 * np.cos(phi)
 
         assert np.allclose(result, expected, tol)
@@ -859,9 +1029,17 @@ class TestExpOperatorArithmetic:
             [qml.expval(qml.prod(qml.PauliZ(0), qml.PauliX(1)))],
         )
         statevector = lightning_sv(self.wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = -np.cos(phi)
 
         assert np.allclose(result, expected, tol)
@@ -874,9 +1052,17 @@ class TestExpOperatorArithmetic:
             [qml.expval(qml.sum(qml.PauliZ(0), qml.PauliX(1)))],
         )
         statevector = lightning_sv(self.wires)
-        statevector = statevector.get_final_state(tape)
+        statevector = (
+            statevector.get_final_state(tape)
+            if device_name != "lightning.tensor"
+            else statevector.set_tensor_network(tape)
+        )
         m = LightningMeasurements(statevector)
-        result = m.measure_final_state(tape)
+        result = (
+            m.measure_final_state(tape)
+            if device_name != "lightning.tensor"
+            else m.measure_tensor_network(tape)
+        )
         expected = np.cos(phi) + np.sin(theta)
 
         assert np.allclose(result, expected, tol)
@@ -900,9 +1086,17 @@ def test_state_vector_2_qubit_subset(tol, op, par, wires, expected, lightning_sv
     )
 
     statevector = lightning_sv(2)
-    statevector = statevector.get_final_state(tape)
+    statevector = (
+        statevector.get_final_state(tape)
+        if device_name != "lightning.tensor"
+        else statevector.set_tensor_network(tape)
+    )
 
     m = LightningMeasurements(statevector)
-    result = m.measure_final_state(tape)
+    result = (
+        m.measure_final_state(tape)
+        if device_name != "lightning.tensor"
+        else m.measure_tensor_network(tape)
+    )
 
     assert np.allclose(result, expected, tol)